refactoring to avoid unwanted driver import
