+++ token/sdk/tms/processor.go
@@ -4,7 +4,7 @@ Copyright IBM Corp. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package token
+package tms
 
 import (
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric"
